requirements.txt
jsonschema
python-qpid-proton
